requirements.txt
flask
aiohttp
requests
python-dotenv